% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{cluster.default}
\alias{cluster.default}
\title{Default algorithm for cluster confidence index}
\usage{
cluster.default(V, nstart = 25, centers = 5)
}
\arguments{
\item{V}{The scGBM scores}

\item{nstart}{A \code{kmeans} parameter}

\item{centers}{A \code{kmeans} parameter}
}
\value{
The function applies \code{kmeans} to \code{V} and returns
the assignments of clusters
}
\description{
Default algorithm for cluster confidence index
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
