% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{denoise.U}
\alias{denoise.U}
\title{Denoise loadings}
\usage{
denoise.U(out)
}
\description{
Use the adaptive shrinkage method implement in \code{ashr}
to stabilize estimates of U.
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
