% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gbm.init}
\alias{gbm.init}
\title{Initialize scGBM fit with PCA on Rank-2 Approximation}
\usage{
gbm.init(Y, M, batch = as.factor(rep(1, ncol(Y))))
}
\arguments{
\item{Y}{data}

\item{M}{rank of approximation}

\item{batch}{batch for each variable}
}
\value{
fit
}
\description{
Initialize scGBM fit with PCA on Rank-2 Approximation
}
