% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings.R
\name{loadings.volcano}
\alias{loadings.volcano}
\title{Plot the factor loadings}
\usage{
loadings.volcano(gbm, dim = 1, return.plot = FALSE)
}
\arguments{
\item{gbm}{A list that is the return value of \code{gbm.sc}}

\item{dim}{Which latent factor to plot}

\item{return.plot}{Should ggplot object containing the plot be returned?}
}
\value{
If `return.plot` is TRUE, a `ggplot2` object is returned.
}
\description{
A volcano plot shows which genes are driving
a particular scGBM factor
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
