% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gbm}
\alias{plot_gbm}
\title{Plot the results of scGBM}
\usage{
plot_gbm(
  gbm,
  dims = c(1, 2),
  cluster = NULL,
  se = FALSE,
  return.gg = FALSE,
  plot = TRUE
)
}
\arguments{
\item{gbm}{An object which is the output of \code{\link{gbm.sc}}}

\item{dims}{Which dimensions (in the latent factors) to plot}

\item{se}{If TRUE, plot the uncertainty estimates around each point}

\item{return.gg}{If TRUE, return the ggplot object}

\item{plot}{If TRUE, plot the \code{ggplot} object}
}
\value{
If \code{return.gg} is TRUE, then a \code{ggplot} object is
returned.
}
\description{
Plot the results of scGBM
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
