% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simData}
\alias{simData}
\title{Simulate data from a Poisson bilinear model}
\usage{
simData(I, J, d, alpha.mean = 0, beta.mean = 0, K = 2)
}
\arguments{
\item{I}{The number of genes}

\item{J}{The number of cells}

\item{d}{The number of latent factors}

\item{alpha.mean}{The mean of the gene-specific intercepts}

\item{beta.mean}{The mean of the cell-specific intercepts}

\item{K}{The latent variability, ratio of the largest to smallest singular value.}
}
\value{
A list with components
\itemize{
\item \code{Y} Simulated count matrix
\item \code{V} Ground truth latent factor scores
\item \code{U} Ground truth latent factor loadings
\item \code{D} A vector of the ground truth scaling factors (singular values)
\item \code{alpha} Ground truth gene-specific intercepts
\item \code{beta} Ground truth cell-specific intercepts
}
}
\description{
Simulate data from a Poisson bilinear model
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
