% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{CCI}
\alias{CCI}
\title{Evaluate cluster stability using the Cluster Confidence Index (CCI)}
\usage{
CCI(gbm, cluster.orig, reps = 100, cluster.fn = cluster.default, ...)
}
\arguments{
\item{gbm}{The scGBM object estimated from \code{gbm.sc}.}

\item{cluster.orig}{A vector of length J containing the labeling of
cells into clusters.}

\item{reps}{The number of simulated datasets to make.}

\item{cluster.fn}{A function that takes as input a set of (simulated) scores V and
returns a labeling of cells into clusters.}
}
\value{
A list with components
\itemize{
\item \code{H.table} - The values used to make the cluster confidence
heatmap.
\item \code{cci_diagonal} - A ggplot object that plots the distribution of intra-cluster
CCI.
\item \code{cci} - A 3-dimensional array of all CCI values (inter-and-intra-cluster).
}
}
\description{
Evaluate cluster stability using the Cluster Confidence Index (CCI)
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
