% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{CCI}
\alias{CCI}
\title{Evaluate cluster stability using the Cluster Confidence Index (CCI)}
\usage{
CCI(gbm, cluster.orig, reps = 100, cluster.fn = cluster.default, ...)
}
\arguments{
\item{gbm}{The scGBM object estimated from \code{gbm.sc}.}

\item{cluster.orig}{A vector of length J containing the labeling of
cells into clusters.}

\item{reps}{The number of simulated datasets to make.}

\item{cluster.fn}{A function that takes as input a set of (simulated) scores V and
returns a labeling of cells into clusters.}
}
\value{

}
\description{
Evaluate cluster stability using the Cluster Confidence Index (CCI)
}
