% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gbm.sc}
\alias{gbm.sc}
\title{scGBM: Model-based dimensionality reduction for single-cell
RNA-seq with generalized bilinear models}
\usage{
gbm.sc(
  Y,
  M,
  max.iter = 100,
  tol = 10^{
     -4
 },
  subset = NULL,
  ncores = 1,
  infer.beta = FALSE,
  return.W = TRUE,
  batch = as.factor(rep(1, ncol(Y))),
  time.by.iter = FALSE,
  min.iter = 30
)
}
\arguments{
\item{Y}{A matrix of UMI counts with genes on rows and cells on columns.}

\item{M}{The number of latent factors to estimate}

\item{max.iter}{The maximum number of iterations}

\item{tol}{The tolerance for convergence (relative difference in objective function)}

\item{subset}{If NULL, use the entire data to fit the model. If an integer, take a
random subsample of cells to fit the loadings. Then project the remaining
cells onto those loadings. If a non-random sample is desired, subset can
be an integer vector corresponding to the columns that should be used to
fit the loadings.}

\item{ncores}{If subset is not NULL, then ncores specifies the number of cores to use
during the projection step.}

\item{infer.beta}{If FALSE, beta is set to the log of the number of counts per cell.}

\item{return.W}{If TRUE, the matrix of weights (which is equal to the estimated mean) is returned. This should
be FALSE for large datasets since it requires a lot of memory.}

\item{batch}{An optional factor containing the assignment of cells to known batches.}

\item{time.by.iter}{If TRUE, the elapsed time (in seconds) is given at each iteration of the algorithm.}

\item{min.iter}{The minimum number of iterations.}
}
\value{
A list with components
\itemize{
\item \code{scores} - A matrix containing the factor scores (weight of each cell for each factor).
\item \code{loadings} - A matrix (genes x factors) containing the factor loadings (weight of each cell for each factor).
\item \code{V} - A matrix containing the unscaled factor scores.
\item \code{U} - Factor loadings. This matrix is identical to \code{loadings} and is included
for reverse compatibility.
\item \code{D} - A vector containing the singular values (scaling factors)
\item \code{alpha} - A matrix containing gene-specific intercepts. The number of columns
is set to be equal to the number of batches (by default there are no batches so this is 1).
\item \code{beta} - A vector of cell-specific intercepts.
\item \code{I} - The number of genes.
\item \code{J} - The number of cells.
\item \code{W} - The estimated mean (and by properties of the Poisson distribution, also the variance)
for each entry of the count matrix.
\item \code{obj} - The value of the objective function for each iteration.
}
}
\description{
Fit a Poisson bilinear model to the single-cell
count matrix.
}
\details{
scGBM fits the following model to the \eqn{I \times J} count matrix \eqn{Y}:
\deqn{Y \sim \text{Poisson}(\mu)}
\deqn{\log \mu = \alpha 1_J^T + 1_I \beta^T + U \Sigma V^T}
Here \eqn{\alpha} are gene-specific intercepts, \eqn{\beta_j} are cell-specific intercepts, \eqn{U} are the \eqn{I \times M} factor loadings
(linear combinations of genes that define factors), and \eqn{V \Sigma} are the \eqn{J \times M} scores for each cell.
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
